datapipeline-gcp/cdc-ingestion/requirements.txt
google-cloud-storage==2.15.0
requests==2.31.0
